switch to a full ts setup
