--- packages/gamut-styles/index.js
@@ -1,1 +1,0 @@
-export * from './utils/variables';
