+++ packages/gamut-styles/package.json
@@ -6,7 +6,9 @@
   "main": "dist/index.js",
   "module": "dist/index.js",
   "scripts": {
-    "build": "tsc",
+    "build:clean": "rm -rf ./dist",
+    "build:compile": "tsc",
+    "build": "yarn build:clean && yarn build:compile",
     "prepublishOnly": "yarn build"
   },
   "repository": {
